Backup dat

+++ QLHS-WebApp/qlhsapp/dao.py
@@ -343,6 +343,30 @@ def update_class(class_id, new_homeroom_teacher_id, class_name):
         return False
 
 
+def add_new_school_year(year1, year2, start_hk1, finish_hk1, start_hk2, finish_hk2):
+    if len(year1) == 4 and len(year2) == 4 and int(year2) == int(year1)+1:
+        start_hk1 = datetime.strptime(start_hk1, "%Y-%m-%d")
+        finish_hk1 = datetime.strptime(finish_hk1, "%Y-%m-%d")
+        start_hk2 = datetime.strptime(start_hk2, "%Y-%m-%d")
+        finish_hk2 = datetime.strptime(finish_hk2, "%Y-%m-%d")
+
+        sy = f'{year1}-{year2}'
+        school_year = SchoolYear(name=sy)
+        db.session.add(school_year)
+        db.session.flush()
+
+        hk1 = Semester(start_date=start_hk1, finish_date=finish_hk1, name='HK1', school_year_id=school_year.id)
+        hk2 = Semester(start_date=start_hk2, finish_date=finish_hk2, name='HK2', school_year_id=school_year.id)
+        db.session.add_all([hk1, hk2])
+        db.session.commit()
+
+        flash('Tạo năm học mới thành công!', 'success')
+        return True
+
+    flash('Năm học phải đủ 4 chữ số và cách nhau 1 năm!', 'danger')
+    return False
+
+
 def add_student_to_class(student_list_id, class_id):
     class_ = Class.query.get(class_id)
     class_max_size = Regulation.query.filter_by(key_name='CLASS_MAX_SIZE').first()
@@ -468,8 +492,6 @@ def add_teaching_assignment(teacher_id, class_id, subject_id, school_year_id):
         exist_ta = TeachingAssignment.query.filter_by(class_id=class_id,
             subject_id=subject_id,school_year_id=school_year_id).first()
 
-        print(teacher_id)
-        print(exist_ta.teacher_id)
         if exist_ta and exist_ta.teacher_id == teacher_id:
             flash(f'Bản phân công này đã tồn tại!!', 'danger')
             return False
@@ -556,14 +578,52 @@ def update_student(student_id, name, address, email, date_of_birth, phone_number
         flash(f"Lỗi update: {str(e)}", "danger")
 
 
+def get_school_years_with_semesters():
+    query = (
+        db.session.query(
+            SchoolYear.id,
+            SchoolYear.name,
+            Semester.name,
+            Semester.start_date,
+            Semester.finish_date
+        )
+        .join(Semester, SchoolYear.id == Semester.school_year_id)
+    )
+    results = query.all()
+    return results
+
+
+def format_school_year_data(results):
+    school_years = {}
+    for school_year_id, school_year_name, semester_name, start_date, finish_date in results:
+        if school_year_id not in school_years:
+            school_years[school_year_id] = {
+                "name": school_year_name,
+                "semesters": {}
+            }
+        school_years[school_year_id]["semesters"][semester_name] = {
+            "start_date": start_date,
+            "finish_date": finish_date
+        }
+    return school_years
+
+
 # Trung code: Xóa học sinh
 def delete_student(student_id):
+    student_from_class = StudentClass.query.filter_by(student_id=student_id, is_active=True).first()
+    class_ = Class.query.get(student_from_class.class_id)
     student = Student.query.get(student_id)
-    if student:
+
+    if student_from_class and student and class_:
         db.session.delete(student)
+        class_.student_numbers -= 1
         db.session.commit()
     else:
         raise ValueError("Không tìm thấy học sinh cần xóa")
+
+
+
+
 
 
 # Trung code: Tiếp nhận học sinh
@@ -636,6 +696,19 @@ def delete_subject(subject_id):
         db.session.commit()
     else:
         raise ValueError("Không tìm thấy môn học cần xóa")
+
+def load_teachers(kw=None, page=1):
+    page_size = app.config['PAGE_SIZE']
+    start = (page - 1) * page_size
+    query = (db.session.query(Teacher, User)
+             .join(Teacher, User.id == Teacher.teacher_id)
+             )
+    total_records = query.count()
+    total_pages = math.ceil(total_records / page_size)
+    if kw:
+        query = query.filter(User.first_name.ilike(f"%{kw}%") | User.last_name.ilike(f"%{kw}%"))
+    results = query.offset(start).limit(page_size).all()
+    return results, total_pages
 
 
 def list_students(kw=None, class_id=None, page=1):
